Rename ir module
